Backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.0.0